pyproject.toml
[project]
name = "reputation-safe-agent"
version = "0.1.0"
description = "A reputation-safe agent blueprint for customer-facing chatbots"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.0.0",
    "pydantic>=2.0.0",
    "qdrant-client>=1.7.0",
    "sentence-transformers>=2.2.0",
    "rich>=13.0.0",
    "tavily-python>=0.3.0",
    "python-dotenv>=1.2.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
